Use initialization function to load _pyodide_core
