+++ src/core/main.c
@@ -1,44 +1,15 @@
 #define PY_SSIZE_T_CLEAN
 #include "Python.h"
 #include <assert.h>
+#include <stdbool.h>
 #include <emscripten.h>
-#include <stdalign.h>
 
-#include "docstring.h"
-#include "error_handling.h"
-#include "hiwire.h"
-#include "js2python.h"
-#include "jsproxy.h"
-#include "pyproxy.h"
-#include "python2js.h"
-#include "python2js_buffer.h"
-
-#define FATAL_ERROR(args...)                                                   \
-  do {                                                                         \
-    EM_ASM(API.fatal_loading_error(args));                                     \
-    return -1;                                                                 \
-  } while (0)
 
 #define FAIL_IF_STATUS_EXCEPTION(status)                                       \
   if (PyStatus_Exception(status)) {                                            \
     goto finally;                                                              \
   }
 
-#define TRY_INIT(mod)                                                          \
-  do {                                                                         \
-    int mod##_init();                                                          \
-    if (mod##_init()) {                                                        \
-      FATAL_ERROR("Failed to initialize module ", #mod, ".");                  \
-    }                                                                          \
-  } while (0)
-
-#define TRY_INIT_WITH_CORE_MODULE(mod)                                         \
-  do {                                                                         \
-    int mod##_init(PyObject* mod);                                             \
-    if (mod##_init(core_module)) {                                             \
-      FATAL_ERROR("Failed to initialize module", #mod, ".");                   \
-    }                                                                          \
-  } while (0)
 
 // Initialize python. exit() and print message to stderr on failure.
 static void
@@ -74,12 +45,8 @@ finally:
   }
 }
 
-static struct PyModuleDef core_module_def = {
-  PyModuleDef_HEAD_INIT,
-  .m_name = "_pyodide_core",
-  .m_doc = "Pyodide C builtins",
-  .m_size = -1,
-};
+PyObject*
+PyInit__pyodide_core(void);
 
 /**
  * Bootstrap steps here:
@@ -96,53 +63,12 @@ main(int argc, char** argv)
 {
   // This exits and prints a message to stderr on failure,
   // no status code to check.
+  PyImport_AppendInittab("_pyodide_core", PyInit__pyodide_core);
   initialize_python(argc, argv);
   emscripten_exit_with_live_runtime();
   return 0;
 }
 
-int
-pyodide_init(void)
-{
-  PyObject* _pyodide = NULL;
-  PyObject* core_module = NULL;
-  JsRef _pyodide_proxy = NULL;
-
-  _pyodide = PyImport_ImportModule("_pyodide");
-  if (_pyodide == NULL) {
-    FATAL_ERROR("Failed to import _pyodide module.");
-  }
-
-  core_module = PyModule_Create(&core_module_def);
-  if (core_module == NULL) {
-    FATAL_ERROR("Failed to create core module.");
-  }
-
-  TRY_INIT_WITH_CORE_MODULE(error_handling);
-  TRY_INIT(hiwire);
-  TRY_INIT(docstring);
-  TRY_INIT(js2python);
-  TRY_INIT_WITH_CORE_MODULE(python2js);
-  TRY_INIT(python2js_buffer);
-  TRY_INIT_WITH_CORE_MODULE(JsProxy);
-  TRY_INIT_WITH_CORE_MODULE(pyproxy);
-
-  PyObject* module_dict = PyImport_GetModuleDict(); /* borrowed */
-  if (PyDict_SetItemString(module_dict, "_pyodide_core", core_module)) {
-    FATAL_ERROR("Failed to add '_pyodide_core' module to modules dict.");
-  }
-
-  // Enable JavaScript access to the _pyodide module.
-  _pyodide_proxy = python2js(_pyodide);
-  if (_pyodide_proxy == NULL) {
-    FATAL_ERROR("Failed to create _pyodide proxy.");
-  }
-  EM_ASM({ API._pyodide = Hiwire.pop_value($0); }, _pyodide_proxy);
-
-  Py_CLEAR(_pyodide);
-  Py_CLEAR(core_module);
-  return 0;
-}
 
 void
 pymain_run_python(int* exitcode);
